added reinforce (no particules)
