--- old/a2c/multi_cpus/main.yaml
@@ -1,28 +1,0 @@
-logger:
-  classname: salina.logger.TFLogger
-  log_dir: ./tmp
-  every_n_seconds: 10
-  verbose: True
-
-algorithm:
-  env_seed: 432
-  n_envs: 4
-  n_processes: 2
-  n_timesteps: 16
-  max_epochs: 10000
-  discount_factor: 0.95
-  entropy_coef: 0.001
-  critic_coef: 1.0
-  a2c_coef: 0.1
-
-  architecture:
-    hidden_size: 32
-
-  env:
-    classname: __main__.make_env
-    env_name: CartPole-v0
-    max_episode_steps: 100
-      
-  optimizer:
-    classname: torch.optim.Adam
-    lr: 0.01
