--- old/reinforce/main.yaml
@@ -1,30 +1,0 @@
-logger:
-  classname: salina.logger.TFLogger
-  log_dir: ./tmp
-  every_n_seconds: 10
-  modulo: 1
-  verbose: True
-
-algorithm:
-  env_seed: 432
-  n_envs: 8
-  n_processes: 1
-  n_particles: 1
-  n_timesteps: 16
-  max_epochs: 500
-  discount_factor: 0.95
-  entropy_coef: 0.001
-  baseline_coef: 1.0
-  reinforce_coef: 0.1
-
-  architecture:
-    hidden_size: 32
-
-  env:
-    classname: __main__.make_env
-    env_name: CartPole-v1
-    max_episode_steps: 500
-
-  optimizer:
-    classname: torch.optim.Adam
-    lr: 0.01
