+++ svpg/algos/reinforce/multi/agents.py
@@ -1,0 +1,33 @@
+import torch
+import torch.nn as nn
+
+from salina import TAgent
+
+
+class REINFORCEAgent(TAgent):
+    def __init__(self, observation_size, hidden_size, n_actions):
+        super().__init__(name="r_agent")
+        self.model = nn.Sequential(
+            nn.Linear(observation_size, hidden_size),
+            nn.ReLU(),
+            nn.Linear(hidden_size, n_actions),
+        )
+        self.critic_model = nn.Sequential(
+            nn.Linear(observation_size, hidden_size),
+            nn.ReLU(),
+            nn.Linear(hidden_size, 1),
+        )
+
+    def forward(self, t, stochastic, **kwargs):
+        observation = self.get(("env/env_obs", t))
+        scores = self.model(observation)
+        probs = torch.softmax(scores, dim=-1)
+        critic = self.critic_model(observation).squeeze(-1)
+        if stochastic:
+            action = torch.distributions.Categorical(probs).sample()
+        else:
+            action = probs.argmax(1)
+
+        self.set(("action", t), action)
+        self.set(("action_probs", t), probs)
+        self.set(("baseline", t), critic)
